2021_legrest copy/requirements.txt
numpy==1.19.5
pandas
matplotlib
notebook
jupyterlab
xlrd
python-dateutil
scikit-learn
scikit-image
scipy
hampel
keras
tensorflow==2.4.1
tensorflow-model-optimization
tensorflow-hub
tensorflow-addons
tensorboard
opencv-python
statsmodels